car_numbers/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
djoser==2.1.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0